+++ basiccheck/string.go
@@ -10,3 +10,10 @@ func StringHasOneOfPrefixes(s string, list []string) bool {
 		return strings.HasPrefix(s, v)
 	})
 }
+
+// StringHasOneOfSuffixes check if string has one of suffix list.
+func StringHasOneOfSuffixes(s string, list []string) bool {
+	return OneInSliceWith(list, func(v string) bool {
+		return strings.HasSuffix(s, v)
+	})
+}